chore(condo): DOMA-5394 reuse @open-condo/keystone/plugins/utils/address-service-client for mock client
