+++ packages/keystone/plugins/utils/address-service-client/MockedAddressServiceClient.js
@@ -2,6 +2,100 @@ const faker = require('faker')
 const get = require('lodash/get')
 
 class MockedAddressServiceClient {
+    DEFAULT_META_DATA = {
+        qc: null,
+        area: null,
+        city: 'Самара',
+        flat: null,
+        block: '1',
+        floor: null,
+        house: '106г',
+        metro: null,
+        okato: '36401385000',
+        oktmo: '36701330',
+        stead: null,
+        qc_geo: '0',
+        region: 'Самарская',
+        source: null,
+        street: 'Ново-Садовая',
+        country: 'Россия',
+        fias_id: '4e232277-2670-427b-bef7-86a8922a2c81',
+        geo_lat: '53.216211',
+        geo_lon: '50.145792',
+        entrance: null,
+        kladr_id: '6300000100007730181',
+        qc_house: null,
+        timezone: null,
+        area_type: null,
+        city_area: null,
+        city_type: 'г',
+        divisions: null,
+        fias_code: null,
+        flat_area: null,
+        flat_type: null,
+        block_type: 'к',
+        fias_level: '8',
+        flat_price: null,
+        geoname_id: 99330,
+        house_type: 'д',
+        postal_box: null,
+        settlement: null,
+        stead_type: null,
+        tax_office: '6316',
+        beltway_hit: null,
+        flat_cadnum: null,
+        postal_code: '443068',
+        qc_complete: null,
+        region_type: 'обл',
+        street_type: 'ул',
+        area_fias_id: null,
+        city_fias_id: 'bb035cc3-1dc2-4627-9d25-a1bf2d4b936b',
+        flat_fias_id: null,
+        house_cadnum: null,
+        stead_cadnum: null,
+        area_kladr_id: null,
+        city_district: 'Октябрьский',
+        city_kladr_id: '6300000100000',
+        house_fias_id: '7c174eb5-e5d7-4532-9ba1-9fcf9148be9b',
+        stead_fias_id: null,
+        area_type_full: null,
+        area_with_type: null,
+        capital_marker: '2',
+        city_type_full: 'город',
+        city_with_type: 'г Самара',
+        flat_type_full: null,
+        history_values: null,
+        house_kladr_id: '6300000100007730181',
+        region_fias_id: 'df3d7359-afa9-4aaa-8ff9-197e73906b1c',
+        street_fias_id: 'ca0fd9f7-225a-49ad-90fa-66153c3704bf',
+        unparsed_parts: null,
+        block_type_full: 'корпус',
+        house_type_full: 'дом',
+        region_iso_code: 'RU-SAM',
+        region_kladr_id: '6300000000000',
+        settlement_type: null,
+        stead_type_full: null,
+        street_kladr_id: '63000001000077300',
+        beltway_distance: null,
+        country_iso_code: 'RU',
+        federal_district: 'Приволжский',
+        region_type_full: 'область',
+        region_with_type: 'Самарская обл',
+        street_type_full: 'улица',
+        street_with_type: 'ул Ново-Садовая',
+        tax_office_legal: '6316',
+        city_district_type: 'р-н',
+        settlement_fias_id: null,
+        square_meter_price: null,
+        settlement_kladr_id: null,
+        fias_actuality_state: '0',
+        settlement_type_full: null,
+        settlement_with_type: null,
+        city_district_fias_id: null,
+        city_district_kladr_id: null,
+        city_district_type_full: 'район',
+        city_district_with_type: 'Октябрьский р-н',
+    }
 
     constructor (existingItem) {
         this.existingItem = existingItem
@@ -15,106 +109,13 @@ class MockedAddressServiceClient {
      */
     async search (s, params = {}) {
         const item = this.existingItem
-        const metaData = {
-            qc: null,
-            area: null,
-            city: 'Самара',
-            flat: null,
-            block: '1',
-            floor: null,
-            house: '106г',
-            metro: null,
-            okato: '36401385000',
-            oktmo: '36701330',
-            stead: null,
-            qc_geo: '0',
-            region: 'Самарская',
-            source: null,
-            street: 'Ново-Садовая',
-            country: 'Россия',
-            fias_id: '4e232277-2670-427b-bef7-86a8922a2c81',
-            geo_lat: '53.216211',
-            geo_lon: '50.145792',
-            entrance: null,
-            kladr_id: '6300000100007730181',
-            qc_house: null,
-            timezone: null,
-            area_type: null,
-            city_area: null,
-            city_type: 'г',
-            divisions: null,
-            fias_code: null,
-            flat_area: null,
-            flat_type: null,
-            block_type: 'к',
-            fias_level: '8',
-            flat_price: null,
-            geoname_id: 99330,
-            house_type: 'д',
-            postal_box: null,
-            settlement: null,
-            stead_type: null,
-            tax_office: '6316',
-            beltway_hit: null,
-            flat_cadnum: null,
-            postal_code: '443068',
-            qc_complete: null,
-            region_type: 'обл',
-            street_type: 'ул',
-            area_fias_id: null,
-            city_fias_id: 'bb035cc3-1dc2-4627-9d25-a1bf2d4b936b',
-            flat_fias_id: null,
-            house_cadnum: null,
-            stead_cadnum: null,
-            area_kladr_id: null,
-            city_district: 'Октябрьский',
-            city_kladr_id: '6300000100000',
-            house_fias_id: '7c174eb5-e5d7-4532-9ba1-9fcf9148be9b',
-            stead_fias_id: null,
-            area_type_full: null,
-            area_with_type: null,
-            capital_marker: '2',
-            city_type_full: 'город',
-            city_with_type: 'г Самара',
-            flat_type_full: null,
-            history_values: null,
-            house_kladr_id: '6300000100007730181',
-            region_fias_id: 'df3d7359-afa9-4aaa-8ff9-197e73906b1c',
-            street_fias_id: 'ca0fd9f7-225a-49ad-90fa-66153c3704bf',
-            unparsed_parts: null,
-            block_type_full: 'корпус',
-            house_type_full: 'дом',
-            region_iso_code: 'RU-SAM',
-            region_kladr_id: '6300000000000',
-            settlement_type: null,
-            stead_type_full: null,
-            street_kladr_id: '63000001000077300',
-            beltway_distance: null,
-            country_iso_code: 'RU',
-            federal_district: 'Приволжский',
-            region_type_full: 'область',
-            region_with_type: 'Самарская обл',
-            street_type_full: 'улица',
-            street_with_type: 'ул Ново-Садовая',
-            tax_office_legal: '6316',
-            city_district_type: 'р-н',
-            settlement_fias_id: null,
-            square_meter_price: null,
-            settlement_kladr_id: null,
-            fias_actuality_state: '0',
-            settlement_type_full: null,
-            settlement_with_type: null,
-            city_district_fias_id: null,
-            city_district_kladr_id: null,
-            city_district_type_full: 'район',
-            city_district_with_type: 'Октябрьский р-н',
-        }
+
         return {
             addressSources: [s],
             address: get(item, 'address', s),
             addressKey: get(item, 'addressKey', faker.random.alphaNumeric(32)),
             addressMeta: {
-                data: get(item, ['addressMeta', 'data'], metaData),
+                data: get(item, ['addressMeta', 'data'], this.DEFAULT_META_DATA),
                 provider: {
                     name: 'dadata',
                     rawData: get(
@@ -123,7 +124,7 @@ class MockedAddressServiceClient {
                         {
                             dv: 1,
                             address: 'г Самара, ул Ново-Садовая, д 106г к 1',
-                            data: metaData,
+                            data: this.DEFAULT_META_DATA,
                             value: 'г Самара, ул Ново-Садовая, д 106г к 1',
                             unrestricted_value: '443068, Самарская обл, г Самара, Октябрьский р-н, ул Ново-Садовая, д 106г к 1',
                         },
@@ -136,7 +137,23 @@ class MockedAddressServiceClient {
     }
 
     async suggest (s, params) {
-        throw new Error('If you really need in this method, realize it! :)')
+        const parts = this.existingItem.address.split(', ')
+        const begin = parts.slice(0, -3)
+        const [cityPart, streetPart, housePart] = parts.slice(-3)
+        let [houseType, houseNumber, ...rest] = (housePart || '').split(' ')
+        houseNumber = parseInt(houseNumber || '1')
+        if (isNaN(houseNumber)) houseNumber = 1
+        const lastPart = [houseType, houseNumber, ...rest].join(' ')
+
+        const value = [...begin, cityPart, streetPart, lastPart].join(', ')
+        const data = JSON.parse(JSON.stringify(this.DEFAULT_META_DATA))
+
+        data['house_type_full'] = houseType === 'д' ? 'дом' : houseType || 'house'
+
+        return [{
+            value,
+            data,
+        }]
     }
 }
